To ENT or not to ENT, that is the question!  I'd rather be with ENT!
Add RFC 8020 Empty Non-Terminal Name support.  This adds a singly linked list
with domain names in the system.  Since there can be multiple domain names
it is a forest.  And we search for ENTS in this forest given a question that
would normally be getting an NXDOMAIN.  Since an NXDOMAIN is not normal
according to RFC 8020 we rewrite output to be NODATA when not dnssec'ed.  When
DNSSEC'ed we return with a NOERROR (this still needs testing).