Change from Berkeley DB to tree(3) macros.
I have for simplicity sake kept the BerkeleyDB db format but had to rename
every variable which was a lot of work.

Add db.c - this is the new delphinusdns DB routines wrapping tree(3).

TLSA was not honored in AXFR, fix that.  That's a major bug for 1.1.0.

Tested on FreeBSD, OpenBSD and Linux Raspbian.