Add parsing for a replicant (aka slave) zone.  The parser understands this:
rzone "centroid.eu." {
        tsigkey "YmFzZTY0Cg==";
        masterport 10053;
        master 127.0.0.1;
        zonename "centroid.eu.";
        filename "/etc/delphinusdns/centroid.eu.repl";
}

and will try to parse the replicant zone.  This was really hard to get right
and I mucked around the YACC lexer for a bit too long tonight.

The configfile version jumps from 8 to 9 now, so if you plan on using this
code from now on bump your version in the configfile(s).