Rework how slave notification works.  We had a list of IP addresses but that was not fine grained enough and we need to pair a zone name with an series of IP addresses (for each slave zone server).  Away goes the notify "" {} config and in goes the mzone "" {} config instead.  An mzone looks like this:
mzone "internal.centroid.eu" {
        zonename "internal.centroid.eu";
        notifydest 192.168.177.40 pass;
        notifydest 192.168.177.1 pass;
}

mzone "words" {
        zonename "words";
        notifydest 192.168.177.1 secret;
        notifydest 192.168.177.21 secret;
}

There also is a (yet untested) notifybind keyword in mzone.  Each zone
has to have an mzone "" {} as well as a zone "" {} for a master setup.

mzone stands for "master zone", but it is an additional option to the zone.

tested on OpenBSD in my network at home against BIND and nsd.