because axfr.c is run mostly in a chroot it cannot access timezone information unless that's also copied into the chroot.  But that's too much overhead for the sysadmin.  So we read out the tm_gmtoff with localtime at start of program before any chroot and save that variable globally.  We then set it back.
This sould fix a bug that I unearthed by resigning a zone at 00:47 in the
morning (not my usual operating hours) and noticed that notifies were not
sent until I reloaded the zone at 01:00, I want it to honor local host time
so this patch.   I will test it later tonight at 00:30 or something.

not tested, but compiles.