conform to RFC 2181 section 5.2.  Basically each RRset has the same TTL..
A very tedious commit, change the ttl on a per RR level to a per RRset level.
An RRSET is all A's, or all AAAA's in a domain name like the apex.  This
caused some confusion with RRSIG records and I had to resort to reading
RFC 4034 again.  With the help of Habbie in #dns I was able to overcome my
confusion.

To quote RFC 4034 section 3:

The TTL value of an RRSIG RR MUST match the TTL value of the RRset it covers.
This is an exception to the [RFC 2181] rules for TTL values of individual RRs
within an RRset;  individual RRSIG RRs with the same owner name will have
different TTL values if the RRsets they cover have different TTL values.

Tested on OpenBSD/amd64