BIG COMMIT TIME - this work pertains to the FORWARD process
- add one log to additional_tsig()
- add dup sockets, these are UDP duplicates and are created in the same order
as the udp sockets, but they do a shutdown(.., SHUT_RD) so they can't read,
this is to prevent them from getting network traffic, they are only for
replying for the FORWARD process.
- remove struct forward from ddd-db.h and move it to struct sforward in ddd-dns.h
- add a tsigoffset in struct tsig, this will help in stripping TSIG RR's from
an authenticated reply.
- add a zonecount variable which counts how many zones we have, if it's 0 a
series of checks get skipped at startup. (such as determine_glue() and init_entlist()
- make sure that forward is an integer and struct forward is called sforward
- in the lookup logic when we reply NO_DATA skip to FORWARD when we were going
to reply REFUSED, becuase we can only be UNDER some internal zones.  We did
goto's in this case forwardudp/forwardtcp.
- make sure we write the forward imsg to the cortex not the cortex control.
- many many changes to forward.c, most significant would likely be
	check_tsig() which is based on build_question(), which skips all RR's
	in a reply and then authenticates the TSIG filling a returning tsig
	struct.
- in util.c lookup_zone logic have some short circuits if we have forwarding
turned on. ie. if zonecount == 0 only forward.