This fixes the following in forward.c:
When a HMAC failed we still cached it apparently this fixes that.
Fix HMAC comparison on retried packets, *hopefully* what happened was that
a retry packet could be in-flight to the dns server but it also sent a reply
at which point the HMAC changed due to time updates.  We now keep a list of
up to 4 HMAC's (5 in the code just in case) and compare this 'maclist' in a
loop of 0 to 4 in check_tsig().
When we retry try to go exponentially higher.  1 second, 4 seconds, 9 seconds,
16 seconds (but at 16 the packet may already be removed).  this will prevent 
unneccessary retries.