change LICENSE from 3 clause BSD license to OpenBSD-compatible license as found at https://cvsweb.openbsd.org/src/share/misc/license.template?rev=HEAD
What changes in particular:

The name of the author (me) may now be used to endors or promote products
derived from this software.  No specific prior written permission is
necessary.

Why this change?
I want to keep my license as short as possible and I like what OpenBSD does
so I'm taking their license and applying it on my code as well.

Do keep the "AS IS" in mind.  I'm not liable in any way of you using my
software.