A weekend of hacking on my laptop (at my parents residence):
fix some unveils and having tested the pledges they seem to work right, though
I seem to think that when there is no rpath, wpath or cpath pledge that an
unveil (ps flag U) is not displayed, unless I'm having problems.

Convert MD5_* functions to EVP_* functions (libressl recommends this in manpage)

Also add a query cache of 10 slots which should speed up delivery of DNS
similarly the log output has changed to reflect a CRC16 to match with a
CACHEHIT, otherwise I don't know what type it is from the logs.
pass sretlen back on UDP sendto's which covers the packet's length that
we sent, which is added to the cache...

The cache so far is a circular queue with no speedups such as trees or 
binary search and no weights.  The last 10 replies get queued serially.
delphinusdnsd does not inspect the packets coming in in logic, all it does is 
it gets the length of the input packet and if that matches a cache it will  
MD5 hash it and if that's a match it will send the pre-computed payload.  
Since DNS is on OSI layer 5 it is oblivious to IP address and port (Although 
I'm going to have to give this greater thought yet).

Some statistics on the cache (from my speediest cloud computer):

Data was gathered since april 1st, 2021 where there was 159414 queries.
The median time to answer from when the packet was tagged with a timestamp as
it was dequeued from the ethernet card to when the sendto system call was
done was 226 microseconds.  Average time to reply a query is 950 microseconds
and fastest answer which btw was replied with a cache size of 4 (instead of 10)
was 24 microseconds.  The highest time to reply to a query, was attributeable 
to the cloud computer having load, was near 200 milliseconds.

Finally I now know the beauty of the original RFC1035 DNS question packet 
which seems was made with caching in mind.  Take the first 2 bytes (ID) off 
and you have a cacheable blob.  Take the last 4 bytes of the blob and you 
have type and class (for accounting), although it seems in retrospect of DNS 
history the EDNS0 OPT headers could have been  disruptive to that and broke 
this elegance.  Too much inspection of the question packet is dangerous as 
it exposes logic to an unknown, which allows possible flaws to make it.