BIND and nsd in a zone transfer turn off the question after the initial TSIG we do the same now in order to stay compatible with last commit: af7851f9c6ef166fc9121171c87fde12dd38872b
This also means that if there is no question, our compression got confused.
So temporarily disable compression in AXFR's.

This whole behaviour can be turned off to the original behaviour (except
the compression) by setting the primary-axfr-old-behaviour option, for
backward compatibilty.

This change was tested between two delphinusdnsd on amd64(primary) and octeon
(replicant).  Update CHANGES.