one small change and a bigger change:
1. I was getting bad refused answers on filters:
	reply_refused hash out zero'ing of additionals

2.  fix up cname's there is a lot of code around this logic and I don't know 
if I hit the right spot.  What I tested now is that it will reply a CNAME to 
any of the supported RR's (just in delphinusdnsd) to the destination 
hostname/RR.

Also I took out code that was broken that was supposed to add on additionals
in reply_cname.

Tested by me on OpenBSD/arm64.