In an effort to find memory corruption, I think I nailed it.  In commit 172454f6d5100b538846eb86ebc459b17da67dc5 (Sep 10, 2021) I committed find_rrsetwild() which was clearly bogus when a \0 query (. or root zone) lookup came in.  I tried to work around this now.
Also fix up the caching which I suspected of causing this corruption.  It
can't hurt to keep the new code in there.

Thanks to those that caused my delphinusdnsd to crash you caused me a long
debug tour on two or three occasions and I think I finally found it!  Thanks
for not killing my daemons completely!