Add CDS and CDNSKEY (RFC 7344) support.  AFAIK these were just DS and DNSKEY records with different RR number, so I was able to copy/paste a lot over and change the RR.
Also found that dddctl query didn't have a bunch of RR support that I added in
the last little while.  Add them and add the date to it in a comment.

Tested with dddctl query which reuses the raxfr_ functions, I still have to
test signing.  So there may be subsequent commits for this.