very experimental code.  In hearsay I heard that other dns forwarders forward to all the forwarders per request, I want to do the same.  This is lightly tested so there may be a chance to revert in a bit.
The way it was before was that each single request would be forwarded to one
of the forwarders not all which is good if there is no packet loss, it just
seemed a bit slow to me though.

The goal here is not to gain speed but to test the other setup.