TLS support, this is highly experimental really leaning on RFC 7858 (somewhat) This code is based on the do53t.c file and modified to work with TLS. In the options to activate this code one must do:
        tls-port 853;
        tls-certfile "/etc/ssl/somehostname.tld.fullchain.pem";
        tls-keyfile "/etc/ssl/private/somehostname.tld.key";

So far it's only working on OpenBSD. tls-port option with portname activates it
to on.

Also dddctl query for a lookup other than an axfr can now lookup with the -t
option, this is the only case where a hostname is required and not an IP (in
the subsequent -Q or -@).

Since it's late I'll sleep on it before activating this.

Tested on OpenBSD/amd64