instead of memset/zero padding the EDNS0 padding, use arc4random_buf() though RFC 7830 had this to say in section 3:
The PADDING octets SHOULD be set to 0x00.  Other values MAY be used,
   for example, in cases where there is a concern that the padded
   message could be subject to compression before encryption.  PADDING
   octets of any value MUST be accepted in the messages received.

so I go for the last line. it MUST be accepted.

I think this is safer than having crypto blocks of NUL-padding which can be
singled out and attempted to be broken (they know the block is all zeros).
So this gives us randomness and more protection against cryptanalysis.