Change the ratelimit algorithm a bit.  We now have a "pointer" to the offsets as an 32 bit integer unsigned.  This is for reasons of 4 byte alignment only.
Next, instead of a time_t which is often 8 bytes on modern UNIX-like OS's...
we reduce this to 16 bits or uint16_t this allows a granularity of 65536
seconds which is nearly a day.  Since we're checking for seconds within the
last second this shouldn't conflict much.

We also introduce a 8 bit pad to get 4 byte alignment into the struct times.
If the 16 bits proves to be too short we can always add the 8 bits as well
from this pad.  But for now, and for sake of speed, keep it not used.

In my tests I have found that with a ratelimit rate of 24 packets per second,
the memory footprints drops from 50 MB to 35 MB.  A savings of 15 MB, this is
worth it for me.

this algorithm is currently in testing, on sky.